+++ src/auth/pages/AuthPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link, Navigate, useLocation, useNavigate } from "react-router-dom";
 import { Mail, Lock, User, Eye, EyeOff, AlertCircle, Loader2 } from "lucide-react";
-import { Navbar } from "../../shared/components/navbar/Navbar";
+import { Navbar } from "../../shared/components/Navbar";
 import Footer from "../../shared/components/Footer";
 import { useTheme } from "../../shared/context/ThemeContext";
 import { useAuth } from "../context/AuthContext";
