+++ src/features/book/Ch1.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { ChevronDown, Cpu, Binary, Calculator, BookOpen, Lightbulb, CheckCircle } from 'lucide-react';
 import {logic_and_computer_design_fundamental} from 'dld-books'
-import { Navbar } from '../../shared/components/navbar/Navbar';
+import { Navbar } from '../../shared/components/Navbar';
 import Footer from '../../shared/components/Footer';
 import { useTheme } from '../../shared/context/ThemeContext';
