+++ src/features/coal/AssemblyDrillsPage.jsx
@@ -14,7 +14,7 @@ import {
   CheckCircle,
   Code
 } from "lucide-react";
-import { Navbar } from "../../shared/components/navbar/Navbar";
+import { Navbar } from "../../shared/components/Navbar";
 import Footer from "../../shared/components/Footer";
 import { useTheme } from "../../shared/context/ThemeContext";
 import usePointerGlow from "../../shared/hooks/usePointerGlow";
